requirements.txt
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.25
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
urllib3==1.26.6
